Don't make firewall changes if unless running
As part of the server task list we make changes to the firewall. As part
of the worker task list we disable the firewall. This means that on a
combined server and worker, runs after the first run will always fail.

We should only attempt to make firewall changes if firewalld is running.